Add parent Fragment to keep common functions.
